chore: Add unit tests for GenAIKeys and related settings
- Implement tests for GenAIKeys factories in `test_factories.py` to ensure proper validation and behavior when required parameters are missing.
- Create comprehensive tests for GenAIKeys functionality in `test_genaikeys.py`, covering secret retrieval, caching, and error handling.
- Introduce logging tests in `test_logging.py` to verify that sensitive information is not logged and that logging behaves as expected.
- Remove outdated tests from `test_secret_keeper.py` to streamline the test suite.
- Add tests for AWS, Azure, and GCP settings in `test_settings.py` to validate environment variable reading and constructor overrides.
